[refactor] getComponentType함수 hypercloud transformUtil로 위치 수정, 기존 transformUtil의 어플리케이션 그룹부분 복구
