+++ frontend/packages/dev-console/src/components/topology/data-transforms/data-transformer.ts
@@ -8,8 +8,8 @@ import { TopologyDataModel, TopologyDataResources, Edge } from '../hypercloud/hy
 import { TYPE_TRAFFIC_CONNECTOR, TYPE_WORKLOAD, TYPE_CONNECTS_TO } from '../components/const';
 import { HelmReleaseResourcesMap } from '../../helm/helm-types';
 import { allowedResources } from '../topology-utils';
-import { addToTopologyDataModel, createInstanceForResource, createTopologyNodeData, getTopologyNodeItem, getComponentType } from './transform-utils';
-import { getChildrenResources } from './hypercloud/transform-utils';
+import { addToTopologyDataModel, createInstanceForResource, createTopologyNodeData, getTopologyNodeItem } from './transform-utils';
+import { getChildrenResources, getComponentType } from './hypercloud/transform-utils';
 // import { getOperatorTopologyDataModel } from '../operators/operators-data-transformer';
 // import { getHelmTopologyDataModel } from '../helm/helm-data-transformer';
